inserting trail with details and photos

+++ app/src/main/java/com/example/trails/model/Coordinates.java
@@ -2,11 +2,11 @@ package com.example.trails.model;
 
 public class Coordinates {
 
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
 
-    public Coordinates(float latitude, float longitude) {
+    public Coordinates(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -14,19 +14,19 @@ public class Coordinates {
     public Coordinates(){
 
     }
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
